split out cdm, bps needs this too
