+++ system/camerad/cameras/camera_common.cc
@@ -114,7 +114,7 @@ bool CameraBuf::acquire(int expo_time) {
 
   double start_time = millis_since_boot();
   //imgproc->runKernel(camera_bufs_raw[cur_buf_idx].buf_cl, cur_yuv_buf->buf_cl, out_img_width, out_img_height, expo_time);
-  memcpy(cur_yuv_buf->addr, cur_camera_buf->addr, cur_camera_buf->len);
+  memcpy(cur_yuv_buf->addr, camera_bufs[cur_buf_idx].addr, camera_bufs[cur_buf_idx].len);
   cur_frame_data.processing_time = (millis_since_boot() - start_time) / 1000.0;
 
   VisionIpcBufExtra extra = {
